Fix tests, add bin/node functionality
Added a source transform in the previous commit which
would add require calls for es6-{promise,map,set} to the
top of each module.

Have also updated es-{map,set} to use GitHub URLs, but
only as a temporary measure.

There's a REPL now too which will transpile code on the fly.